ref(chat): Split bot.ts into composition and runtime services
Extract bot wiring from orchestration, state, conversation memory, and vision logic into focused runtime and service modules. This keeps behavior unchanged while making ownership boundaries explicit and improving testability and navigability for Slack bot development.
