+++ src/chat/bootstrap/register-handlers.ts
@@ -1,0 +1,76 @@
+import type { Chat } from "chat";
+import type { SlackAdapter } from "@chat-adapter/slack";
+import type { AppRuntimeAssistantLifecycleEvent, AppSlackRuntime } from "@/chat/app-runtime";
+import { getUserTokenStore } from "@/chat/capabilities/factory";
+import { logException, withSpan } from "@/chat/observability";
+import { publishAppHomeView } from "@/chat/app-home";
+import { handleSlashCommand } from "@/chat/slash-command";
+import { getSlackClient } from "@/chat/slack-actions/client";
+
+export function registerBotHandlers(args: {
+  bot: Chat<{ slack: SlackAdapter }>;
+  appSlackRuntime: AppSlackRuntime<unknown, AppRuntimeAssistantLifecycleEvent>;
+}): void {
+  const { bot, appSlackRuntime } = args;
+
+  bot.onNewMention(appSlackRuntime.handleNewMention);
+  bot.onSubscribedMessage(appSlackRuntime.handleSubscribedMessage);
+  bot.onAssistantThreadStarted((event: AppRuntimeAssistantLifecycleEvent) =>
+    appSlackRuntime.handleAssistantThreadStarted(event)
+  );
+  bot.onAssistantContextChanged((event: AppRuntimeAssistantLifecycleEvent) =>
+    appSlackRuntime.handleAssistantContextChanged(event)
+  );
+
+  bot.onSlashCommand("/jr", (event) =>
+    withSpan(
+      "workflow.slash_command",
+      "workflow.slash_command",
+      { slackUserId: event.user.userId },
+      async () => {
+        try {
+          await handleSlashCommand(event);
+        } catch (error) {
+          logException(error, "slash_command_failed", { slackUserId: event.user.userId });
+          throw error;
+        }
+      }
+    )
+  );
+
+  bot.onAppHomeOpened((event) =>
+    withSpan(
+      "workflow.app_home_opened",
+      "workflow.app_home_opened",
+      { slackUserId: event.userId },
+      async () => {
+        try {
+          await publishAppHomeView(getSlackClient(), event.userId, getUserTokenStore());
+        } catch (error) {
+          logException(error, "app_home_opened_failed", { slackUserId: event.userId });
+        }
+      }
+    )
+  );
+
+  bot.onAction("app_home_disconnect", async (event) => {
+    const provider = event.value;
+    if (!provider) return;
+    const userId = event.user.userId;
+    await withSpan(
+      "workflow.app_home_disconnect",
+      "workflow.app_home_disconnect",
+      { slackUserId: userId },
+      async () => {
+        try {
+          await getUserTokenStore().delete(userId, provider);
+          await publishAppHomeView(getSlackClient(), userId, getUserTokenStore());
+        } catch (error) {
+          logException(error, "app_home_disconnect_failed", { slackUserId: userId }, {
+            "app.credential.provider": provider
+          });
+        }
+      }
+    );
+  });
+}
